Minor refactoring
- move FeatureMetadata.Feature to a separate type
