+++ core/runtime/src/main/java/io/quarkiverse/mcp/server/runtime/FeatureMetadata.java
@@ -40,17 +40,4 @@ public record FeatureMetadata<M>(Feature feature, FeatureMethodInfo info, Invoke
         return json;
     }
 
-    public enum Feature {
-        PROMPT,
-        TOOL,
-        RESOURCE,
-        RESOURCE_TEMPLATE,
-        PROMPT_COMPLETE,
-        RESOURCE_TEMPLATE_COMPLETE;
-
-        public boolean requiresUri() {
-            return this == RESOURCE || this == RESOURCE_TEMPLATE;
-        }
-    }
-
 }
